Do not reuse dns.Envelope variable
Sending a reference to a dns.Envelope (&env) to a channel and then
resetting the env variable to a fresh envelope causes the reference sent
over the channel to also reference an empty envelope.

Now we maintain a separate slice of dns.RR records and create a new
envelope on each send instead. Passing the dns.RR slice over the channel
is OK as even if it internally points to data, the slice header itself
will be copied and is not affected when we reset "rrs" to a new empty
slice.